integration-tests/requirements.txt
grpcio>=1.63.0
grpcio-tools>=1.63.0
PyJWT>=2.8.0
pytest>=8.2.0
pytest-timeout>=2.3.1
testcontainers>=4.8.0
requests>=2.31.0
docker>=7.0.0
